requirements.txt
asgiref==3.6.0
bleach==6.0.0
Django==4.1.5
djangorestframework==3.14.0
drf-api-logger==1.1.12
Pillow==9.4.0
PyMySQL==1.0.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
webencodings==0.5.1

requests~=2.28.2